+++ src/executor/styles.rs
@@ -2037,3 +2037,81 @@ impl Region {
         }
     }
 }
+
+#[derive(serde::Deserialize, Debug, PartialEq, Clone, serde::Serialize)]
+pub enum WhiteSpace {
+    NORMAL,
+    NOWRAP,
+    PRE,
+    PREWRAP,
+    PRELINE,
+    BREAKSPACES,
+}
+
+impl WhiteSpace {
+    fn from_optional_values(
+        or_type_value: Option<(String, ftd::interpreter2::PropertyValue)>,
+        doc: &ftd::executor::TDoc,
+        line_number: usize,
+    ) -> ftd::executor::Result<Option<Self>> {
+        if let Some(value) = or_type_value {
+            Ok(Some(WhiteSpace::from_values(value, doc, line_number)?))
+        } else {
+            Ok(None)
+        }
+    }
+
+    fn from_values(
+        or_type_value: (String, ftd::interpreter2::PropertyValue),
+        doc: &ftd::executor::TDoc,
+        line_number: usize,
+    ) -> ftd::executor::Result<Self> {
+        match or_type_value.0.as_str() {
+            ftd::interpreter2::FTD_WHITESPACE_NORMAL => Ok(WhiteSpace::NORMAL),
+            ftd::interpreter2::FTD_WHITESPACE_NOWRAP => Ok(WhiteSpace::NOWRAP),
+            ftd::interpreter2::FTD_WHITESPACE_PRE => Ok(WhiteSpace::PRE),
+            ftd::interpreter2::FTD_WHITESPACE_PREWRAP => Ok(WhiteSpace::PREWRAP),
+            ftd::interpreter2::FTD_WHITESPACE_PRELINE => Ok(WhiteSpace::PRELINE),
+            ftd::interpreter2::FTD_WHITESPACE_BREAKSPACES => Ok(WhiteSpace::BREAKSPACES),
+            t => ftd::executor::utils::parse_error(
+                format!("Unknown variant `{}` for or-type `ftd.whitespace`", t),
+                doc.name,
+                line_number,
+            ),
+        }
+    }
+
+    pub(crate) fn optional_whitespace(
+        properties: &[ftd::interpreter2::Property],
+        arguments: &[ftd::interpreter2::Argument],
+        doc: &ftd::executor::TDoc,
+        line_number: usize,
+        key: &str,
+    ) -> ftd::executor::Result<ftd::executor::Value<Option<WhiteSpace>>> {
+        let or_type_value = ftd::executor::value::optional_or_type(
+            key,
+            properties,
+            arguments,
+            doc,
+            line_number,
+            ftd::interpreter2::FTD_WHITESPACE,
+        )?;
+
+        Ok(ftd::executor::Value::new(
+            WhiteSpace::from_optional_values(or_type_value.value, doc, line_number)?,
+            or_type_value.line_number,
+            or_type_value.properties,
+        ))
+    }
+
+    pub fn to_css_string(&self) -> String {
+        match self {
+            WhiteSpace::NORMAL => "normal".to_string(),
+            WhiteSpace::NOWRAP => "nowrap".to_string(),
+            WhiteSpace::PRE => "pre".to_string(),
+            WhiteSpace::PRELINE => "pre-line".to_string(),
+            WhiteSpace::PREWRAP => "pre-wrap".to_string(),
+            WhiteSpace::BREAKSPACES => "break-spaces".to_string(),
+        }
+    }
+}